### 2021-06-23 a2e494a
Added filter properties and a weight for TravelRouteSetEntries within a TravelRouteSet
