+++ Database/Tables/Transportation/TravelRouteSetEntry.cs
@@ -2,6 +2,7 @@
 using System.Collections.ObjectModel;
 using System.Linq;
 using Automation;
+using Common.Enums;
 using Database.Database;
 using JetBrains.Annotations;
 
@@ -12,9 +13,14 @@ namespace Database.Tables.Transportation {
         [CanBeNull] private readonly TravelRoute _travelRoute;
 
         private readonly int _travelRouteSetID;
+        public readonly int _minimumAge;
+        public readonly int _maximumAge;
+        public readonly PermittedGender _gender;
+        public readonly int _affordanceTaggingSetEntryID;
+        public readonly double _weight;
 
-        public TravelRouteSetEntry([CanBeNull]int? pID, int travelRouteSetID, [JetBrains.Annotations.NotNull] string connectionString, [JetBrains.Annotations.NotNull] string name,
-            [CanBeNull] TravelRoute travelRoute, [NotNull] StrGuid guid)
+    public TravelRouteSetEntry([CanBeNull]int? pID, int travelRouteSetID, [JetBrains.Annotations.NotNull] string connectionString, [JetBrains.Annotations.NotNull] string name,
+            [CanBeNull] TravelRoute travelRoute, int minimumAge, int maximumAge, PermittedGender gender, int affordanceTaggingSetEntryID, double weight, [NotNull] StrGuid guid)
             : base(name, TableName, connectionString, guid)
         {
             TypeDescription = "Travel Route Step";
@@ -22,6 +28,11 @@ namespace Database.Tables.Transportation {
 
             _travelRoute = travelRoute;
             _travelRouteSetID = travelRouteSetID;
+            _minimumAge = minimumAge;
+            _maximumAge = maximumAge;
+            _gender = gender;
+            _affordanceTaggingSetEntryID = affordanceTaggingSetEntryID;
+            _weight = weight;
         }
 
         [UsedImplicitly]
@@ -31,7 +42,22 @@ namespace Database.Tables.Transportation {
         [UsedImplicitly]
         public int TravelRouteSetID => _travelRouteSetID;
 
-        [JetBrains.Annotations.NotNull]
+        [UsedImplicitly]
+        public int MinimumAge => _minimumAge;
+
+        [UsedImplicitly]
+        public int MaximumAge => _maximumAge;
+
+        [UsedImplicitly]
+        public PermittedGender Gender => _gender;
+
+        [UsedImplicitly]
+        public int AffordanceTaggingSetEntryID => _affordanceTaggingSetEntryID;
+
+        [UsedImplicitly]
+        public double Weight => _weight;
+
+    [JetBrains.Annotations.NotNull]
         private static TravelRouteSetEntry AssignFields([JetBrains.Annotations.NotNull] DataReader dr, [JetBrains.Annotations.NotNull] string connectionString,
             bool ignoreMissingFields,
             [JetBrains.Annotations.NotNull] AllItemCollections aic)
@@ -40,10 +66,15 @@ namespace Database.Tables.Transportation {
             var setid = dr.GetIntFromLong("TravelRouteSetID");
             var routeID = dr.GetIntFromLong("TravelRouteID");
             var route = aic.TravelRoutes.FirstOrDefault(x => x.ID == routeID);
+            var minimumAge = dr.GetIntFromLong("MinimumAge",false, ignoreMissingFields,-1);
+            var maximumAge = dr.GetIntFromLong("MaximumAge", false, ignoreMissingFields, -1);
+            var gender = (PermittedGender) dr.GetIntFromLong("Gender", false, ignoreMissingFields, -1);
+            var affordanceTaggingSetEntryID = dr.GetIntFromLong("AffordanceTaggingSetEntryID", false, ignoreMissingFields, -1);
[…]
