### 2021-06-23 787e305
Added another possible path to search for the DB file in the project directory
