### 2021-06-30 ed4c10e
Adapted TravelRouteSetEntry to contain an AffordanceTag instead of an AffordanceTaggingSetEntryID
