+++ SimulationEngineLib/HouseJobProcessor/HouseGenerator.cs
@@ -566,7 +566,7 @@ namespace SimulationEngineLib.HouseJobProcessor {
                 }
 
                 if (addUnmodifiedRoute) {
-                    adjustedTravelrouteset.AddRoute(oldTravelRouteSetEntry.TravelRoute, oldTravelRouteSetEntry.MinimumAge, oldTravelRouteSetEntry.MaximumAge, oldTravelRouteSetEntry.Gender, oldTravelRouteSetEntry.AffordanceTaggingSetEntryID, oldTravelRouteSetEntry.Weight);
+                    adjustedTravelrouteset.AddRoute(oldTravelRouteSetEntry.TravelRoute, oldTravelRouteSetEntry.MinimumAge, oldTravelRouteSetEntry.MaximumAge, oldTravelRouteSetEntry.Gender, oldTravelRouteSetEntry.AffordanceTag, oldTravelRouteSetEntry.Weight);
                 }
             }
             //Config.ShowDeleteMessages = true;
@@ -606,7 +606,7 @@ namespace SimulationEngineLib.HouseJobProcessor {
 
             newRoute.SaveToDB();
             //Logger.Info("Adjusted route " + newRoute.Name);
-            adjustedTravelrouteset.AddRoute(newRoute, oldTravelRouteSetEntry.MinimumAge, oldTravelRouteSetEntry.MaximumAge, oldTravelRouteSetEntry.Gender, oldTravelRouteSetEntry.AffordanceTaggingSetEntryID, oldTravelRouteSetEntry.Weight);
+            adjustedTravelrouteset.AddRoute(newRoute, oldTravelRouteSetEntry.MinimumAge, oldTravelRouteSetEntry.MaximumAge, oldTravelRouteSetEntry.Gender, oldTravelRouteSetEntry.AffordanceTag, oldTravelRouteSetEntry.Weight);
         }
 
         public static void WriteGuidList([NotNull] string filename, [NotNull] [ItemNotNull] List<DBBase> elements, [NotNull] DirectoryInfo relativePath)
