### 2021-07-05 e535c3c
Added filter options to CalcTravelRoute; added affordance-tag dicts to transportation handler
