+++ CalculationController/CalcFactories/CalcManagerFactory.cs
@@ -121,7 +121,7 @@ namespace CalculationController.CalcFactories
                 csps.CalculationProfiler.StartPart(Utili.GetCurrentMethodAndClass() + " Generating Model");
                 var container = builder.Build();
                 using var scope = container.BeginLifetimeScope();
-                var calcRepo = PrepareCalculation(sim, csps, scope, out var dtoltdict, out var dls, out var variableRepository);
+                var calcRepo = PrepareCalculation(sim, csps, scope, out var dtoltdict, out var dls, out var variableRepository, out var affordanceTaggingSets);
 
                 cm = new CalcManager(csps.ResultPath,
                     //hh.Name,
@@ -141,7 +141,7 @@ namespace CalculationController.CalcFactories
                 CalcVariableDtoFactory cvrdto = scope.Resolve<CalcVariableDtoFactory>();
                 CalcDeviceTaggingSets devicetaggingSets = scope.Resolve<CalcDeviceTaggingSets>();
                 if (csps.CalcTarget.GetType() == typeof(House)) {
-                    ch = MakeCalcHouseObject(sim, csps, csps.CalcTarget, scope, cvrdto, variableRepository, out cot, calcRepo);
+                    ch = MakeCalcHouseObject(sim, csps, csps.CalcTarget, scope, cvrdto, variableRepository, affordanceTaggingSets, out cot, calcRepo);
                     CalcHouse chd = (CalcHouse)ch;
                     if (chd.EnergyStorages != null) {
                         foreach (var calcEnergyStorage in chd.EnergyStorages) {
@@ -153,7 +153,7 @@ namespace CalculationController.CalcFactories
 
                 }
                 else if (csps.CalcTarget.GetType() == typeof(ModularHousehold)) {
-                    ch = MakeCalcHouseholdObject(sim, csps, csps.CalcTarget, scope, cvrdto, variableRepository, out cot, calcRepo);
+                    ch = MakeCalcHouseholdObject(sim, csps, csps.CalcTarget, scope, cvrdto, variableRepository, out cot, affordanceTaggingSets, calcRepo);
                 }
                 else {
                     throw new LPGException("The type " + csps.CalcTarget.GetType() + " is missing!");
@@ -192,7 +192,8 @@ namespace CalculationController.CalcFactories
                                                    [JetBrains.Annotations.NotNull] ILifetimeScope scope,
                                                    [JetBrains.Annotations.NotNull] out CalcLoadTypeDtoDictionary dtoltdict,
                                                    [JetBrains.Annotations.NotNull] out DayLightStatus dls,
-                                                   [JetBrains.Annotations.NotNull] out CalcVariableRepository variableRepository
+                                                   [JetBrains.Annotations.NotNull] out CalcVariableRepository variableRepository,
+                                                   [JetBrains.Annotations.NotNull] out List<CalcAffordanceTaggingSetDto> affordanceTaggingSets
                                                    )
         {
             CalcRepo calcRepo = scope.Resolve<CalcRepo>();
@@ -203,7 +204,7 @@ namespace CalculationController.CalcFactories
             dtoltdict = scope.Resolve<CalcLoadTypeDtoDictionary>();
 
             var affordanceTaggingSetFactory = scope.Resolve<AffordanceTaggingSetFactory>();
-            var affordanceTaggingSets = affordanceTaggingSetFactory.GetAffordanceTaggingSets(sim);
+            affordanceTaggingSets = affordanceTaggingSetFactory.GetAffordanceTaggingSets(sim);
             if (calcRepo.CalcParameters.Options.Contains(CalcOption.AffordanceTaggingSets)) {
                 inputDataLogger.Save(affordanceTaggingSets);
[…]

[…]
