### 2021-07-05 bc40a85
Implemented filter for matching tag names of affordances and TravelRoutes
