### 2021-07-05 291e469
Adapted tests to the changes regarding TravelRoutes and TravelRouteSets
