+++ Database.Tests/Tables/Transportation/TravelRouteSetTests.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections.ObjectModel;
 using Automation;
 using Common;
+using Common.Enums;
 using Common.Tests;
 using Database.Tables.Transportation;
 using FluentAssertions;
@@ -22,7 +23,7 @@ namespace Database.Tests.Tables.Transportation
                 db.ClearTable(TravelRouteSet.TableName);
                 db.ClearTable(TravelRouteSetEntry.TableName);
 
-                TravelRouteSet set = new TravelRouteSet("set1", null, db.ConnectionString, "desc", System.Guid.NewGuid().ToStrGuid());
+                TravelRouteSet set = new TravelRouteSet("set1", null, db.ConnectionString, "desc", System.Guid.NewGuid().ToStrGuid(), null);
                 set.SaveToDB();
                 Site a = new Site("a", null, db.ConnectionString, "desc", System.Guid.NewGuid().ToStrGuid());
                 a.SaveToDB();
@@ -30,14 +31,20 @@ namespace Database.Tests.Tables.Transportation
                 b.SaveToDB();
                 TravelRoute route = new TravelRoute(null, db.ConnectionString, "routename", "routedesc", a, b, System.Guid.NewGuid().ToStrGuid(), null);
                 route.SaveToDB();
-                set.AddRoute(route);
+                int minimumAge = 10;
+                int maximumAge = 50;
+                PermittedGender gender = PermittedGender.Male;
+                double weighting = 1.0;
+                set.AddRoute(route, minimumAge, maximumAge, gender, null, weighting);
                 //loading
                 ObservableCollection<TravelRoute> routes = new ObservableCollection<TravelRoute>
             {
                 route
             };
                 ObservableCollection<TravelRouteSet> sets = new ObservableCollection<TravelRouteSet>();
-                TravelRouteSet.LoadFromDatabase(sets, db.ConnectionString, false, routes);
+                var affordances = db.LoadAffordances(out _, out _, out _, out _, out _, out var loadTypes, out _, out _, out _, out _, out _ , out _);
+                var affordanceTaggingSets = db.LoadAffordanceTaggingSets(affordances, loadTypes);
+                TravelRouteSet.LoadFromDatabase(sets, db.ConnectionString, false, routes, affordanceTaggingSets);
                 db.Cleanup();
                 (sets.Count).Should().Be(1);
             }
