+++ CalculationController.Tests/CalcFactories/CalcTransportationFactoryTests.cs
@@ -49,7 +49,7 @@ namespace CalculationController.Tests.CalcFactories {
         [Trait(UnitTestCategories.Category,UnitTestCategories.BasicTest)]
         public void CheckRouteCompletenessTest()
         {
-            TravelRouteSet trs = new TravelRouteSet("trs", 1, "", "", Guid.NewGuid().ToStrGuid());
+            TravelRouteSet trs = new TravelRouteSet("trs", 1, "", "", Guid.NewGuid().ToStrGuid(), null);
             List<Site> sites = new List<Site>();
             //test with empty
             CalcTransportationDtoFactory.CheckRouteCompleteness(trs,sites);
@@ -59,17 +59,17 @@ namespace CalculationController.Tests.CalcFactories {
             Site siteb = new Site("siteb", 2, "", "", Guid.NewGuid().ToStrGuid());
             sites.Add(sitea);
             sites.Add(siteb);
-            trs.AddRoute(new TravelRoute(1,"","Route1","desc",sitea,siteb, Guid.NewGuid().ToStrGuid(),null),false);
+            trs.AddRoute(new TravelRoute(1,"","Route1","desc",sitea,siteb, Guid.NewGuid().ToStrGuid(),null), 0, 80, Common.Enums.PermittedGender.All, null, 1.0, false);
             CalcTransportationDtoFactory.CheckRouteCompleteness(trs, sites);
 
             //one route missing
             Site sitec = new Site("sitec", 2, "", "", Guid.NewGuid().ToStrGuid());
             sites.Add(sitec);
-            trs.AddRoute(new TravelRoute(1, "", "Route2", "desc", siteb, sitec, Guid.NewGuid().ToStrGuid(),null), false);
+            trs.AddRoute(new TravelRoute(1, "", "Route2", "desc", siteb, sitec, Guid.NewGuid().ToStrGuid(),null), 0, 80, Common.Enums.PermittedGender.All, null, 1.0, false);
             Action  crashfunction1 = () => CalcTransportationDtoFactory.CheckRouteCompleteness(trs, sites);
             crashfunction1.Should().Throw<DataIntegrityException>();
             //add missing route
-            trs.AddRoute(new TravelRoute(1, "", "Route3", "desc", sitea, sitec, Guid.NewGuid().ToStrGuid(),null), false);
+            trs.AddRoute(new TravelRoute(1, "", "Route3", "desc", sitea, sitec, Guid.NewGuid().ToStrGuid(),null), 0, 80, Common.Enums.PermittedGender.All, null, 1.0, false);
             CalcTransportationDtoFactory.CheckRouteCompleteness(trs, sites);
         }
 

[…]
