+++ Calculation.Tests/Transportation/AffordanceBaseTransportDecoratorTests.cs
@@ -35,22 +35,22 @@ namespace Calculation.Tests.Transportation
                 wd.InputDataLogger.AddSaver(new HouseholdKeyLogger(wd.SqlResultLoggingService));
                 CalcParameters calcParameters = CalcParametersFactory.MakeGoodDefaults();
                 HouseholdKey key = new HouseholdKey("hh1");
-                using (var calcRepo = SetupFullWorkingTransportationExample(wd, rnd, out _, out var srcloc, out var dstloc, out var dstSite, out var transportationHandler, out var abt, calcParameters, key))
+                using (var calcRepo = SetupFullWorkingTransportationExample(wd, rnd, out _, out var srcloc, out var dstloc, out var srcSite, out var dstSite, out var sourceAff, out var transportationHandler, out var abt, calcParameters, key))
                 {
                     //make sure there is a travel route
-                    const string personname = "activator";
                     TimeStep ts = new TimeStep(0, 0, false);
                     var affs = dstloc.Affordances.ToList();
                     var aff = affs[0];
+                    var person = new CalcPersonDto("activator", null, 30, PermittedGender.All, null, null, null, -1, null, null);
                     var travelroute = transportationHandler.GetTravelRouteFromSrcLoc(srcloc, dstSite,
-                        ts, personname, aff, calcRepo);
+                        ts, person, sourceAff, calcRepo);
                     Assert.NotNull(travelroute);
                     // find if busy
-                    var isbusy = abt.IsBusy(ts, srcloc, personname, false);
+                    var isbusy = abt.IsBusy(ts, srcloc, person, false);
                     isbusy.Should().Be(BusynessType.NotBusy);
 
                     Logger.Info("Activating affordance for time 0");
-                    aff.Activate(ts, personname, srcloc, out _);
+                    aff.Activate(ts, person.Name, srcloc, out _);
                 }
                 //should throw exception the second time.
                 Logger.Info("Activating affordance again for time 0");
@@ -72,22 +72,24 @@ namespace Calculation.Tests.Transportation
                 var rnd = new Random(1);
                 CalcParameters calcParameters = CalcParametersFactory.MakeGoodDefaults();
                 HouseholdKey key = new HouseholdKey("hh1");
-                using (var calcRepo = SetupFullWorkingTransportationExample(wd, rnd, out _, out var srcloc, out var dstloc, out var dstSite, out var transportationHandler, out var abt, calcParameters, key))
+                using (var calcRepo = SetupFullWorkingTransportationExample(wd, rnd, out _, out var srcloc, out var dstloc, out var srcSite,
+                    out var dstSite, out var sourceAff, out var transportationHandler, out var abt, calcParameters, key))
                 {
                     //make sure there is a travel route
-                    const string personName = "activator";
                     TimeStep ts = new TimeStep(0, 0, false);
                     var affs = dstloc.Affordances.ToList();
                     var aff = affs[0];
+                    var person = new CalcPersonDto("activator", null, 30, PermittedGender.All, null, null, null, -1, null, null);
+
                     var travelroute = transportationHandler.GetTravelRouteFromSrcLoc(srcloc,
-                        dstSite, ts, personName, aff, calcRepo);
+                        dstSite, ts, person, sourceAff, calcRepo);
                     Assert.NotNull(travelroute);
                     // find if busy
-                    var isbusy = abt.IsBusy(ts, srcloc, "", false);
[…]
