### 2021-07-14 b8ac884
Assigned default route in route selection (only for e.g. floating point errors)
