### 2021-07-15 #3341 @kthyng
Hi @fraukewiese and welcome to JOSS! Does this submission look like a reasonable fit for you to edit?
