### 2021-07-20 c3da29d
Added ownership model for transport devices so that e.g. cars are not left at inappropriate sites
