### 2021-08-03 #3341 @fraukewiese
@brynpickering, @mewilhel, @samuelduchesne
– would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: [https://joss.readthedocs.io/en/latest/reviewer_guidelines.html](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html)

The submission under review here is [LoadProfileGenerator: An Agent-Based Behavior Simulation for Generating Residential Load Profiles](https://github.com/FZJ-IEK3-VSA/LoadProfileGenerator).
