### 2021-08-04 #3341 @mewilhel
Hi,

Looks interesting. I'd be happy to review this.

Best,
Matt

On Tue, Aug 3, 2021 at 10:06 AM fraukewiese ***@***.***>
wrote:

> *Message sent from a system outside of UConn.*
>
> @brynpickering <https://github.com/brynpickering>, @mewilhel
> <https://github.com/mewilhel>, @samuelduchesne
> <https://github.com/samuelduchesne>
> – would any of you be willing to review this submission for JOSS? We carry
> out our checklist-driven reviews here in GitHub issues and follow these
> guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
>
> The submission under review here is LoadProfileGenerator: An Agent-Based
> Behavior Simulation for Generating Residential Load Profiles
> <https://github.com/FZJ-IEK3-VSA/LoadProfileGenerator>.
>
