### 2021-08-05 #3574 @whedon
[REVIEW]: LoadProfileGenerator: An Agent-Based Behavior Simulation for Generating Residential Load Profiles
**Submitting author:** @noah80 (<a href="http://orcid.org/0000-0002-1982-8794">Noah</a>)
**Repository:** https://github.com/FZJ-IEK3-VSA/LoadProfileGenerator
**Branch with paper.md** (empty if default branch):
**Version:** v10.8.0
