### 2021-08-10 5c28fbf
Extended the JSON import to include transport preferences of persons. Instead of JsonReference objects it now is sufficient to provide only the name string.
