+++ SimulationEngineLib/PythonGenerator.cs
@@ -69,6 +69,8 @@ namespace SimulationEngineLib
                 WriteJsonRefs(sim.ChargingStationSets.Items.Select(x => (DBBase)x).ToList(), sw, "ChargingStationSets");
                 WriteJsonRefs(sim.TravelRouteSets.Items.Select(x => (DBBase)x).ToList(), sw, "TravelRouteSets");
                 WriteJsonRefs(sim.Houses.Items.Select(x => (DBBase)x).ToList(), sw, "Houses");
+                WriteJsonRefs(sim.Sites.Items.Select(x => (DBBase)x).ToList(), sw, "Sites");
+                WriteJsonRefs(sim.TransportationDeviceCategories.Items.Select(x => (DBBase)x).ToList(), sw, "TransportationDeviceCategories");
             WriteNames(sim.HouseholdTags.Items.Select(x => (DBBase)x).ToList(), sw, "HouseholdTags");
             WriteNames(sim.LivingPatternTags.Items.Select(x => (DBBase)x).ToList(), sw, "LivingPatternTags");
             WriteNames(sim.HouseholdTemplates.Items.Select(x => (DBBase)x).ToList(), sw, "HouseholdTemplates");
@@ -178,6 +180,7 @@ namespace SimulationEngineLib
                 WriteEnum<HouseholdKeyType>(sw, writtenTypes);
             var encounteredTypes = new List<string>();
                 WriteClass<StrGuid>(sw, encounteredTypes, writtenTypes);
+                WriteClass<TransportationPreference>(sw, encounteredTypes, writtenTypes);
                 WriteClass<PersonData>(sw, encounteredTypes, writtenTypes);
                 WriteClass<JsonReference>(sw, encounteredTypes, writtenTypes);
                 WriteClass<TransportationDistanceModifier>(sw, encounteredTypes, writtenTypes);
@@ -297,10 +300,20 @@ namespace SimulationEngineLib
                     typename = "List[PersonLivingTag]";
                     return info.Name + ": List[PersonLivingTag] = field(default_factory=list)";
                 }
-            if (fulltypename.StartsWith("System.Collections.Generic.List`1[[Automation.PersonData,"))
+                if (fulltypename.StartsWith("System.Collections.Generic.List`1[[Automation.PersonData,"))
                 {
                     typename = "List[PersonData]";
                     return info.Name + ": List[PersonData] = field(default_factory=list)";
+                }
+                if (fulltypename.StartsWith("System.Collections.Generic.List`1[[Automation.TransportationPreference"))
+                {
+                    typename = "List[TransportationPreference]";
+                    return info.Name + ": List[TransportationPreference] = field(default_factory=list)";
+                }
+                if (fulltypename.StartsWith("System.Collections.Generic.List`1[[Automation.JsonReference"))
+                {
+                    typename = "List[JsonReference]";
+                    return info.Name + ": List[JsonReference] = field(default_factory=list)";
                 }
                 if (fulltypename.StartsWith("System.Collections.Generic.List`1[[Automation.TransportationDistanceModifier, "))
                 {
@@ -311,13 +324,7 @@ namespace SimulationEngineLib
                 {
                     typename = "List[HouseholdData]";
                     return info.Name + ": List[HouseholdData] = field(default_factory=list)";
-            }
-                if (fulltypename.StartsWith("System.Collections.Generic.List`1[[System.Double,"))
-                {
-                    typename = "List[float]";
-                    return info.Name + ": List[float] = field(default_factory=list)";
                 }
-
                 if (fulltypename.StartsWith("System.Collections.Generic.List`1[[System.Double,"))
                 {
                     typename = "List[float]";
@@ -329,7 +336,7 @@ namespace SimulationEngineLib
[…]
