+++ Automation/PersonData.cs
@@ -1,12 +1,18 @@
 ﻿
 
+using System.Collections.Generic;
+
 namespace Automation {
-    public class PersonData {
-        public PersonData(int age, Gender gender, string personName)
+    public class PersonData
+    {
+        public PersonData(int age, Gender gender, string personName) : this(age, gender, personName, null) { }
+
+        public PersonData(int age, Gender gender, string personName, List<TransportationPreference>? transportationPreferences)
         {
             Age = age;
             Gender = gender;
             PersonName = personName;
+            TransportationPreferences = transportationPreferences;
         }
 
         public int Age { get; set; }
@@ -14,5 +20,7 @@ namespace Automation {
 
         public string? LivingPatternTag { get; set; }
         public string PersonName { get; set; }
+
+        public List<TransportationPreference>? TransportationPreferences { get; set; }
     }
 }
