### 2021-08-12 0d0d833
Added missing properties for TransportPreferences of a person
