### 2021-08-12 6719080
Minor bugfixes (from failing unit tests)
