+++ SimulationEngineLib/HouseJobProcessor/HouseGenerator.cs
@@ -494,6 +494,12 @@ namespace SimulationEngineLib.HouseJobProcessor {
                     throw new LPGPBadParameterException("Could not find charging station set: " + householdData.ChargingStationSet?.ToString());
                 }
                 var travelrouteset = sim.TravelRouteSets.FindByJsonReference(householdData.TravelRouteSet);
+                if (travelrouteset == null)
+                {
+                    // Alternative transport specification: for each person transportations preferences can be specified. These are used
+                    // to create a new TravelRoueSet.
+                    travelrouteset = CreateTravelRouteSetFromPersonPreferences(householdData, sim);
+                }
                 if (hj.CalcSpec.EnableTransportation && travelrouteset == null)
                 {
                     throw new LPGPBadParameterException("Could not find travel route set.");
@@ -530,6 +536,146 @@ namespace SimulationEngineLib.HouseJobProcessor {
                 throw new LPGException("House generation failed");
             }
             return house.GetJsonReference();
+        }
+
+        /// <summary>
+        /// Alternatively to directly specifying a TravelRouteSet the transport preferences of each person can be specified.
+        /// This function collects these preferences and uses them to create a new TravelRouteSet.
+        /// </summary>
+        /// <param name="householdData">The HouseholdData object for which the TravelRouteSet will be created</param>
+        /// <param name="sim"></param>
+        /// <returns>A new TravelRouteSet that contains all required routes.</returns>
+        private static TravelRouteSet CreateTravelRouteSetFromPersonPreferences(HouseholdData householdData, Simulator sim)
+        {
+            var name = "Generated TravelRouteSet " + "(" + householdData.Name + ")";
+            var description = "This TravelRouteSet was generated using the transportation device preferences of all persons in this household.";
+            var travelRouteSet = new TravelRouteSet(name, null, sim.ConnectionString, description, Guid.NewGuid().ToStrGuid(), null);
+            travelRouteSet.SaveToDB();
+            // find the home site of this household because it has a special role when using transportation preferences of persons
+            Site home = sim.Sites.FindFirstByNameNotNull("Home", FindMode.IgnoreCase);
+            foreach (PersonData person in householdData.HouseholdDataPersonSpec.Persons)
+            {
+                CreateTravelRoutesForPerson(person, home, travelRouteSet, sim);
+            }
+            return travelRouteSet;
+        }
+
+        /// <summary>
+        /// Creates new TravelRoutes based on the transportation preferences of a person and adds them to the TravelRouteSet.
+        /// </summary>
+        /// <param name="person">The person the TravelRoutes will be created for</param>
+        /// <param name="home">The site that is used as home of the person</param>
+        /// <param name="travelRouteSet">The TravelRouteSet to which the new routes will be added</param>
+        /// <param name="sim"></param>
+        private static void CreateTravelRoutesForPerson(PersonData person, Site home, TravelRouteSet travelRouteSet, Simulator sim)
+        {
+            var preferences = person.TransportationPreferences;
+            if (preferences == null)
+            {
+                return;
+            }
+            Dictionary<Site, TransportationPreference> sites = new Dictionary<Site, TransportationPreference>();
+            // collect all sites (except from home)
+            foreach (var preference in preferences)
+            {
+                var site = sim.Sites.FindByJsonReference(preference.DestinationSite);
+                if (site == null)
+                {
+                    throw new LPGPBadParameterException("Could not find the site \"" + site.Name + "\".");
+                }
+                sites.Add(site, preference);
+            }
+            // create routes from each site to all others
+            foreach (var destination in sites.Keys)
+            {
+                var preference = sites[destination];
+                // add route from home
[…]
