### 2021-08-23 c7ef882
Removed unused PDF folder (entire code was commented out)
