### 2021-08-23 #3574 @mewilhel
@fraukewiese I've looked over the software and paper and I'm ready to go through my checklist but I'm receiving the following message when I attempt to accept the invite: "Sorry, we couldn't find that repository invitation. It is possible that the invitation was revoked or that you are not logged into the invited account." What are good next steps to resolve this?
