### 2021-08-26 e8898ad
Added enum to either filter TravelRoutes by person or by age and gender
