### 2021-08-30 27279d1
Affordance tagging set, person and affordance tag of a TravelRouteSet can now be deselected again
