+++ WpfApplication1/Views/Transportation/TravelRouteSetView.xaml.cs
@@ -1,5 +1,6 @@
 ﻿using System.Diagnostics.CodeAnalysis;
 using System.Windows;
+using System.Windows.Controls;
 using System.Windows.Input;
 using Common;
 using Common.Enums;
@@ -7,28 +8,32 @@ using Database.Tables.BasicElements;
 using Database.Tables.Transportation;
 using JetBrains.Annotations;
 
-namespace LoadProfileGenerator.Views.Transportation {
+namespace LoadProfileGenerator.Views.Transportation
+{
     /// <summary>
     ///     Interaktionslogik für DeviceActionView.xaml
     /// </summary>
     [SuppressMessage("ReSharper", "NotNullMemberIsNotInitialized")]
-    public partial class TravelRouteSetView {
+    public partial class TravelRouteSetView
+    {
         public TravelRouteSetView()
         {
             InitializeComponent();
         }
 
         [JetBrains.Annotations.NotNull]
-        private TravelRouteSetPresenter Presenter => (TravelRouteSetPresenter) DataContext;
+        private TravelRouteSetPresenter Presenter => (TravelRouteSetPresenter)DataContext;
 
         private void BtnAddDeviceClick([CanBeNull] object sender, [CanBeNull] RoutedEventArgs e)
         {
-            if (CmbTravelRoutes.SelectedItem == null) {
+            if (CmbTravelRoutes.SelectedItem == null)
+            {
                 Logger.Error("Please select a travel route first!");
                 return;
             }
 
-            if (!(CmbTravelRoutes.SelectedItem is TravelRoute tp)) {
+            if (!(CmbTravelRoutes.SelectedItem is TravelRoute tp))
+            {
                 Logger.Error("Please select a travel route first!");
                 return;
             }
@@ -78,7 +83,8 @@ namespace LoadProfileGenerator.Views.Transportation {
 
         private void BtnRemoveRoutesClick([CanBeNull] object sender, [CanBeNull] RoutedEventArgs e)
         {
-            if (!(LstSteps.SelectedItem is TravelRouteSetEntry entry)) {
+            if (!(LstSteps.SelectedItem is TravelRouteSetEntry entry))
+            {
                 Logger.Error("Please select a TravelRoute entry first!");
                 return;
             }
@@ -95,21 +101,23 @@ namespace LoadProfileGenerator.Views.Transportation {
 
         private void LstPersonDesiresMouseDoubleClick([CanBeNull] object sender, [CanBeNull] MouseButtonEventArgs e)
         {
-            if (LstPersonDesires.SelectedItem == null) {
+            if (LstPersonDesires.SelectedItem == null)
+            {
                 return;
             }
 
-            var ui = (UsedIn) LstPersonDesires.SelectedItem;
+            var ui = (UsedIn)LstPersonDesires.SelectedItem;
             Presenter.ApplicationPresenter.OpenItem(ui.Item);
         }
 
         private void LstSteps_OnMouseDoubleClick([JetBrains.Annotations.NotNull] object sender, [JetBrains.Annotations.NotNull] MouseButtonEventArgs e)
         {
-            if (LstSteps.SelectedItem == null) {
+            if (LstSteps.SelectedItem == null)
+            {
                 return;
             }
 
-            var entry = (TravelRouteSetEntry) LstSteps.SelectedItem;
+            var entry = (TravelRouteSetEntry)LstSteps.SelectedItem;
             Presenter.ApplicationPresenter.OpenItem(entry.TravelRoute);
         }
 
@@ -133,6 +141,38 @@ namespace LoadProfileGenerator.Views.Transportation {
         {
             Presenter.AddDistanceMatchingWorkplaceRoutes();
             Presenter.RefreshRoutes();
[…]
