### 2021-08-31 7a1cbc1
Adapted some hard-coded paths in order to build a new lpg release version
