+++ VersionInfo.cs
@@ -1,3 +1,3 @@
-[assembly: AssemblyVersion("10.5.0.59")]
+[assembly: AssemblyVersion("10.6.0.68")]
 
-[assembly: AssemblyFileVersion("10.5.0.59")]
+[assembly: AssemblyFileVersion("10.6.0.68")]
