+++ ReleaseMaker/SimEngine2Copier.cs
@@ -7,294 +7,17 @@ namespace ReleaseMaker {
         {
             List<string> programFiles = new List<string>();
             var srcDi = new DirectoryInfo(src);
-            Copy(programFiles, srcDi, src, dst, @"api-ms-win-core-console-l1-1-0.dll");
-            Copy(programFiles, srcDi, src, dst, @"api-ms-win-core-console-l1-2-0.dll");
-            Copy(programFiles, srcDi, src, dst, @"api-ms-win-core-datetime-l1-1-0.dll");
-            Copy(programFiles, srcDi, src, dst, @"api-ms-win-core-debug-l1-1-0.dll");
-            Copy(programFiles, srcDi, src, dst, @"api-ms-win-core-errorhandling-l1-1-0.dll");
-            Copy(programFiles, srcDi, src, dst, @"api-ms-win-core-file-l1-1-0.dll");
-            Copy(programFiles, srcDi, src, dst, @"api-ms-win-core-file-l1-2-0.dll");
-            Copy(programFiles, srcDi, src, dst, @"api-ms-win-core-file-l2-1-0.dll");
-            Copy(programFiles, srcDi, src, dst, @"api-ms-win-core-handle-l1-1-0.dll");
-            Copy(programFiles, srcDi, src, dst, @"api-ms-win-core-heap-l1-1-0.dll");
-            Copy(programFiles, srcDi, src, dst, @"api-ms-win-core-interlocked-l1-1-0.dll");
-            Copy(programFiles, srcDi, src, dst, @"api-ms-win-core-libraryloader-l1-1-0.dll");
-            Copy(programFiles, srcDi, src, dst, @"api-ms-win-core-localization-l1-2-0.dll");
-            Copy(programFiles, srcDi, src, dst, @"api-ms-win-core-memory-l1-1-0.dll");
-            Copy(programFiles, srcDi, src, dst, @"api-ms-win-core-namedpipe-l1-1-0.dll");
-            Copy(programFiles, srcDi, src, dst, @"api-ms-win-core-processenvironment-l1-1-0.dll");
-            Copy(programFiles, srcDi, src, dst, @"api-ms-win-core-processthreads-l1-1-0.dll");
-            Copy(programFiles, srcDi, src, dst, @"api-ms-win-core-processthreads-l1-1-1.dll");
-            Copy(programFiles, srcDi, src, dst, @"api-ms-win-core-profile-l1-1-0.dll");
-            Copy(programFiles, srcDi, src, dst, @"api-ms-win-core-rtlsupport-l1-1-0.dll");
-            Copy(programFiles, srcDi, src, dst, @"api-ms-win-core-string-l1-1-0.dll");
-            Copy(programFiles, srcDi, src, dst, @"api-ms-win-core-synch-l1-1-0.dll");
-            Copy(programFiles, srcDi, src, dst, @"api-ms-win-core-synch-l1-2-0.dll");
-            Copy(programFiles, srcDi, src, dst, @"api-ms-win-core-sysinfo-l1-1-0.dll");
-            Copy(programFiles, srcDi, src, dst, @"api-ms-win-core-timezone-l1-1-0.dll");
-            Copy(programFiles, srcDi, src, dst, @"api-ms-win-core-util-l1-1-0.dll");
-            Copy(programFiles, srcDi, src, dst, @"api-ms-win-crt-conio-l1-1-0.dll");
-            Copy(programFiles, srcDi, src, dst, @"api-ms-win-crt-convert-l1-1-0.dll");
-            Copy(programFiles, srcDi, src, dst, @"api-ms-win-crt-environment-l1-1-0.dll");
[…]
