### 2021-09-14 #3574 @mewilhel
Alright, I’ve finished reviewing.

I had an initial issue when attempting to use the software based on system time zones. I was able to easily work around this. I’m an American reviewer and this package appear to be more targeted for an EU end-user so I don’t consider this issue to be gating but rather a chance for continuous improvement (error log text below).

Otherwise, everything loads well on Windows and Linux. The software addresses an existing gap and I found the functionality easy to use with the existing documentation. The paper is well-written and as such I recommend it for publication in its current state.

_errorlog.txt_
An error occurred.  The error message is:
An exception was thrown while activating λ:CalculationEngine.Helper.DayLightStatus.
The time zone set in Windows is not compatible with the time zone of the geographic location. Yes, this is stupid, but it's a bug in the library used to calculate the sunrise times.
The LPG version is 10.5.0.59.
The error was logged to the file errorlog.txt. Please send that file to the author of the program to support further development.
