### 2021-09-14 1800bd2
Added integrity checks for new travel route restrictions
