### 2021-09-14 #3574 @noah80
Thank you very much for the kind review!

The time zone error message only shows up if the time zone of the computer doesn't match the time zone you want to simulate. So if you want to select a german household, you need to set your computer to the german time zone. This is usually not an issue, since most people need load profiles for their home country. It's a bug in the library, as the error message says.

The sunrise/sunset-calculations can be used to determine when people turn on the light, but tests showed that using measured global solar irradiation data works better anyway. People tend to turn on the light in the winter for example if it is too dark even if the sun is up. So, the entire functionality is a bit obsolete and will eventually be removed.
