### 2021-09-17 #3574 @fraukewiese
@szkafander @sptennak
Would one of you be willing to review the submission [LoadProfileGenerator: An Agent-Based Behavior Simulation for Generating Residential Load Profiles](https://github.com/FZJ-IEK3-VSA/LoadProfileGenerator) ?
The review has started already, but another reviewer could not make it, so it would be great if you could review this submission, since your programming skills with C# fit perfectly and also the content might be of interest to you.
