### 2021-10-05 8e8de02
Fixed copying of the lpg folder when exporting a settlement as json calculation definition
