### 2021-10-06 e773871
Checked if the selected view actually contains a grid
