### 2021-10-11 b7a9dea
Views without a stackpanel (e.g. WelcomeView) are now ignored instead of raising an exception.
