### 2021-10-12 c1e5433
Added try-finally to ensure Profiler.StopPart is called
