### 2021-10-12 d4e98c5
Added DatFileDeletor as PostProcessingStep and removed all old references to it
