### 2021-10-12 087a014
Refactored CalcStartParameterSet to remove obsolete deleteDatFiles argument
