+++ CalculationController.Tests/PostProcessingTests.cs
@@ -62,10 +62,10 @@ namespace CalculationController.Tests
                     CalculationProfiler calculationProfiler = new CalculationProfiler();
                     CalcStartParameterSet csps = new CalcStartParameterSet(sim.GeographicLocations[0],
                         sim.TemperatureProfiles[0], sim.ModularHouseholds[0], EnergyIntensityType.Random,
-                        false,  null, LoadTypePriority.All, null, null, null,
+                        false, null, LoadTypePriority.All, null, null, null,
                         sim.MyGeneralConfig.AllEnabledOptions(), new DateTime(2018, 1, 1), new DateTime(2018, 1, 2), new TimeSpan(0, 1, 0),
-                        ";", 5, new TimeSpan(0, 10, 0), false, false, false, 3, 3,
-                        calculationProfiler, wd1.WorkingDirectory,false,false, ".",false);
+                        ";", 5, new TimeSpan(0, 10, 0), false, false, 3, 3, calculationProfiler,
+                        wd1.WorkingDirectory, false, false, ".", false);
                     var cm = cmf.GetCalcManager(sim,  csps, false);
 
                     static bool ReportCancelFunc1()
@@ -144,9 +144,9 @@ namespace CalculationController.Tests
                     options.Add(CalcOption.PolysunImportFiles);
                     CalcStartParameterSet csps = new CalcStartParameterSet(sim.GeographicLocations[0],
                         sim.TemperatureProfiles[0], sim.ModularHouseholds[0], EnergyIntensityType.Random,
-                        false,  null, LoadTypePriority.All, null, null, null, options ,
-                        new DateTime(2013, 1, 1), new DateTime(2013, 1, 2), new TimeSpan(0, 1, 0), ";", 5, new TimeSpan(0, 10, 0), false, false, false, 3, 3,
-                        calculationProfiler, wd1.WorkingDirectory,false,false, decimalSep, false);
+                        false, null, LoadTypePriority.All, null, null, null, options,
+                        new DateTime(2013, 1, 1), new DateTime(2013, 1, 2), new TimeSpan(0, 1, 0), ";", 5, new TimeSpan(0, 10, 0), false, false, 3, 3, calculationProfiler,
+                        wd1.WorkingDirectory, false, false, decimalSep, false);
                     var cm = cmf.GetCalcManager(sim, csps, false);
                     cm.Run(ReportCancelFunc);
                     Logger.ImportantInfo("Duration:" + (DateTime.Now - start).TotalSeconds + " seconds");
