+++ ChartCreator2.Tests/Oxyplot/OxyCalculationSetup.cs
@@ -136,11 +136,11 @@ namespace ChartCreator2.Tests.Oxyplot
 
                 var cspsHouse = new CalcStartParameterSet(ReportFinishFuncForHouseAndSettlement,
                     ReportFinishFuncForHousehold, OpenTabFunc, null, sim.GeographicLocations[0],
-                    sim.TemperatureProfiles[0], house,  energyIntensity, ReportCancelFunc, false,
-                    null, priority, null, null,sim.MyGeneralConfig.AllEnabledOptions(),
-                    sim.MyGeneralConfig.StartDateDateTime,sim.MyGeneralConfig.EndDateDateTime,sim.MyGeneralConfig.InternalStepSize,
-                    ";",-1,new TimeSpan(0,15,0),false,false,false,3, 3,calculationProfiler,null,null,
-                    DeviceProfileHeaderMode.Standard,false, workingDir,false,false, ".", false);
+                    sim.TemperatureProfiles[0], house, energyIntensity, ReportCancelFunc, false,
+                    null, priority, null, null, sim.MyGeneralConfig.AllEnabledOptions(),
+                    sim.MyGeneralConfig.StartDateDateTime, sim.MyGeneralConfig.EndDateDateTime, sim.MyGeneralConfig.InternalStepSize,
+                    ";", -1, new TimeSpan(0, 15, 0), false, false, 3, 3, calculationProfiler, null, null, DeviceProfileHeaderMode.Standard,
+                    false, workingDir, false, false, ".", false);
                 var duration = cspsHouse.OfficialSimulationEndTime - cspsHouse.OfficialSimulationStartTime;
                 if (duration.TotalDays > 370) {
                     throw new LPGException("Trying to test with more than 1 year");
@@ -156,8 +156,8 @@ namespace ChartCreator2.Tests.Oxyplot
                 sim.TemperatureProfiles[0], chh, energyIntensity, ReportCancelFunc, false,
                  null, priority, null, null, sim.MyGeneralConfig.AllEnabledOptions(),
                 sim.MyGeneralConfig.StartDateDateTime, sim.MyGeneralConfig.EndDateDateTime, sim.MyGeneralConfig.InternalStepSize,
-                ";", -1, new TimeSpan(0, 15, 0),false,false,false,3,3,calculationProfiler,null,null,
-                 DeviceProfileHeaderMode.Standard,false, workingDir,false,false, ".", false);
+                ";", -1, new TimeSpan(0, 15, 0), false, false, 3, 3, calculationProfiler, null, null, DeviceProfileHeaderMode.Standard,
+                 false, workingDir, false, false, ".", false);
             var simduration = csps.OfficialSimulationEndTime - csps.OfficialSimulationStartTime;
             if (simduration.TotalDays > 370)
             {
