### 2021-10-14 3c98e1c
Implemented the DatFileDeletor as a post processing step
