### 2021-10-15 0c41e57
Removed old DeleteDatFiles option from SettingsView
