### 2021-10-19 #3574 @noah80
Thanks for the review!

- Build instructions: Good idea. I'll add those.
- Downloadable Manual: I'll add a link.
- Python Wrapper:  Good point. I'll add that.
- Default Path: It's in the settings.
- Time Zone Bug: Will be fixed soon.

I'll update this thread as soon as there is a new version. Hope that this is ok?
