### 2021-10-20 d715964
In AffordanceTaggingSets special Idleness affordances are now intercepted and treated
