### 2021-10-20 1484d65
Fixed generation of CalcOptionTests, and generated these tests anew
