### 2021-10-20 1581715
Renamed helper function
