+++ VersionInfo.cs
@@ -1,3 +1,3 @@
-[assembly: AssemblyVersion("10.6.0.70")]
+[assembly: AssemblyVersion("10.7.0.0")]
 
-[assembly: AssemblyFileVersion("10.6.0.70")]
+[assembly: AssemblyFileVersion("10.7.0.0")]
