### 2021-10-27 43de5d8
Brought back a removed property of AffordanceTaggingSetDto that is needed for json deserialization
