+++ VersionInfo.cs
@@ -1,3 +1,3 @@
-[assembly: AssemblyVersion("10.7.0.0")]
+[assembly: AssemblyVersion("10.7.0.5")]
 
-[assembly: AssemblyFileVersion("10.7.0.0")]
+[assembly: AssemblyFileVersion("10.7.0.5")]
