### 2021-11-11 f5c6b04
Replaced sunlight calculation with solar radiation profiles and a radiation threshold
