### 2021-11-11 bece8fc
Replaced hard coded number with property for default value
