### 2021-11-11 075f422
Removed obsolete classes for sunrise calculation
