+++ Database/Helpers/CalcSunriseTimes.cs
@@ -1,431 +1,0 @@
-﻿////-----------------------------------------------------------------------
-
-//// <copyright>
-////
-//// Copyright (c) TU Chemnitz, Prof. Technische Thermodynamik
-//// Written by Noah Pflugradt.
-//// Redistribution and use in source and binary forms, with or without modification, are permitted provided that the following conditions are met:
-////
-//// Redistributions of source code must retain the above copyright notice, this list of conditions and the following disclaimer.
-////  Redistributions in binary form must reproduce the above copyright notice, this list of conditions and the following disclaimer
-//// in the documentation and/or other materials provided with the distribution.
-////  All advertising materials mentioning features or use of this software must display the following acknowledgement:
-////  “This product includes software developed by the TU Chemnitz, Prof. Technische Thermodynamik and its contributors.”
-////  Neither the name of the University nor the names of its contributors may be used to endorse or promote products
-////  derived from this software without specific prior written permission.
-////
-//// THIS SOFTWARE IS PROVIDED BY THE UNIVERSITY 'AS IS' AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING,
-//// BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
-//// DISCLAIMED. IN NO EVENT SHALL THE UNIVERSITY OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, S
-//// PECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; L
-//// OSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT,
-//// STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF
-//// ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-
-//// </copyright>
-
-////-----------------------------------------------------------------------
-
-//#region
-
-//using System;
-//using System.Diagnostics.CodeAnalysis;
-//using Automation.ResultFiles;
-//using Common;
-
-//#endregion
-
-//namespace Database.Helpers {
-////////////////////////////////////////////////////////////////////////////////////////////////////////
-////
-////  C# Singleton class and thread-safe class for calculating Sunrise and Sunset times.
-////
-//// The algorithm was adapted from the JavaScript sample provided here:
-////      http://home.att.net/~srschmitt/script_sun_rise_set.html
-////
-////  NOTICE: this code is provided "as-is", without any warrenty, obligations or liability for it.
-////          You may use this code freely for any use.
-////
-////  Zacky Pickholz (zacky.pickholz@gmail.com)
-////
-///////////////////////////////////////////////////////////////////////////////////////////////////////
-
-//    public sealed class CalcSunriseTimes {
-//        private const double Mk1 = 15 * PiBy180 * 1.0027379;
-//        private const double PiBy180 = Math.PI / 180;
-//        [JetBrains.Annotations.NotNull] private readonly double[] _mDecensionArr = {0.0, 0.0, 0.0};
-//        [JetBrains.Annotations.NotNull] private readonly object _mLock = new object();
[…]
