### 2021-11-16 efab424
Fixed the exception about the missing db3-file incorrectly occuring for commands where another path to the DB is available
