### 2021-11-17 2814ae0
Converted an obsolete sunrise test into a solar radiation test for all geographic locations
