### 2021-11-23 b3c4fca
Included the filename in the error message about filenames with more than 260 characteres
