### 2021-11-24 c23a8ff
Fixed a test checking for exceptions if the db3 file is missing
