### 2021-11-24 ac0ce5f
C:\work is now used as working directory for tests by default to avoid long path issues
