### 2021-11-29 a004d05
Fixed initialization of datatables for TravelRouteSet preview
