### 2021-11-29 1312e1d
Added necessary CalcOption in plot test to avoid missing table in database
