+++ ChartCreator2.Tests/Oxyplot/DeviceProfilesTests.cs
@@ -25,6 +25,7 @@ namespace ChartCreator2.Tests.Oxyplot {
                 {
                     x.Enable(CalcOption.DeviceProfilesIndividualHouseholds);
                     x.Enable(CalcOption.HouseholdContents);
+                    x.Enable(CalcOption.DeviceTaggingSets);
                     x.SelectedEnergyIntensity = EnergyIntensityType.EnergySavingPreferMeasured;
                     x.EndDateDateTime = new DateTime(2012, 12, 31);
                 }, energyIntensity: EnergyIntensityType.EnergySavingPreferMeasured);
