### 2021-12-06 aaee131
Added a link to the LPG manual to the readme and to the WelcomeView of the LPG
