### 2021-12-13 9b14f3c
Added household template tests for automatic generation and simulation of households from templates
