+++ SimulationEngine.Tests/CalculationTests.cs
@@ -34,6 +34,7 @@ using Xunit.Abstractions;
 
 namespace SimulationEngine.Tests {
     public enum TestDuration {
+        ThreeDays,
         OneMonth,
         ThreeMonths,
         TwelveMonths
@@ -52,6 +53,7 @@ namespace SimulationEngine.Tests {
 
         public static HouseCreationAndCalculationJob PrepareNewHouseForHouseholdTesting(Simulator sim, string guid, TestDuration duration)
         {
+
             var hj = new HouseCreationAndCalculationJob
             {
                 CalcSpec = JsonCalcSpecification.MakeDefaultsForTesting()
@@ -86,17 +88,22 @@ namespace SimulationEngine.Tests {
             if (hj.CalcSpec == null) {
                 throw new LPGException("was null");
             }
-            if (duration == TestDuration.OneMonth) {
-                hj.CalcSpec.EndDate = new DateTime(2020, 2, 1);
-            }
-            else if (duration == TestDuration.ThreeMonths) {
-                hj.CalcSpec.EndDate = new DateTime(2020, 4, 1);
-            }
-            else if (duration == TestDuration.TwelveMonths) {
-                hj.CalcSpec.EndDate = new DateTime(2020, 12, 31);
-            }
-            else {
-                throw new LPGException("Unkown duration");
+            switch (duration)
+            {
+                case TestDuration.ThreeDays:
+                    hj.CalcSpec.EndDate = new DateTime(2020, 1, 4);
+                    break;
+                case TestDuration.OneMonth:
+                    hj.CalcSpec.EndDate = new DateTime(2020, 2, 1);
+                    break;
+                case TestDuration.ThreeMonths:
+                    hj.CalcSpec.EndDate = new DateTime(2020, 4, 1);
+                    break;
+                case TestDuration.TwelveMonths:
+                    hj.CalcSpec.EndDate = new DateTime(2020, 12, 31);
+                    break;
+                default:
+                    throw new LPGException("Unkown duration");
             }
         }
 
@@ -215,7 +222,7 @@ namespace SimulationEngine.Tests {
         {
 
              var peakWorkingSet = Process.GetCurrentProcess().PeakWorkingSet64;
-             const long memoryCap = 1024l * 1024l * 2000l * 2l;
+             const long memoryCap = 1024L * 1024L * 2000L * 2L;
             peakWorkingSet.Should().BeLessThan(memoryCap);
             GC.Collect();
             GC.WaitForPendingFinalizers();
@@ -370,7 +377,7 @@ namespace SimulationEngine.Tests {
             //}
         }
 
-        public static void RunSingleHouse(Func<Simulator, HouseCreationAndCalculationJob> makeHj, Action<string> checkResults,
+        public static void RunSingleHouse(Func<Simulator, HouseCreationAndCalculationJob?> makeHj, Action<string> checkResults,
                                           bool skipcleaning = false)
         {
             Logger.Get().StartCollectingAllMessages();
@@ -384,13 +391,47 @@ namespace SimulationEngine.Tests {
             var sim = new Simulator(db.ConnectionString);
             var houseGenerator = new HouseGenerator();
             var hj = makeHj(sim);
+            hj.Should().NotBeNull();
             houseGenerator.ProcessSingleHouseJob(hj, sim);
-            if (hj.CalcSpec?.OutputDirectory == null)
+            if (hj!.CalcSpec?.OutputDirectory == null)
             {
                 throw new LPGException("calcspec was null");
             }
 
             checkResults(wd.Combine(hj.CalcSpec.OutputDirectory));
+        }
+
+        /// <summary>
+        /// Generates one household from a household template and simulates them.
+        /// </summary>
[…]
