### 2021-12-14 aa25e2d
Adapted generation of python bindings: take pylpg package into account for imports
