### 2021-12-14 6e6910d
Adapted TemplateTests to include template name
