+++ SimulationEngine.Tests/CalculationTests.cs
@@ -859,14 +859,15 @@ namespace SimulationEngine.Tests {
         /// template, and simulates each household for 3 days.
         /// </summary>
         /// <param name="hhTemplateGuid">Guid of the template to test</param>
-        /// <param name="number">Repetition index of this test for the same template guid</param>
+        /// <param name="hhTemplateName">Name of the template to test</param>
+        /// <param name="repetition">Repetition of this test for the same template</param>
         /// <remarks>Each call of this method only handles a single household</remarks>
         [Theory]
         [MemberData(nameof(GetEachHHTemplate100Times))]
         [Trait(UnitTestCategories.Category, UnitTestCategories.HouseholdTemplateTest)]
-        public void SystematicHouseholdTemplateTestShort(StrGuid hhTemplateGuid, int number = 0)
+        public void SystematicHouseholdTemplateTestShort(StrGuid hhTemplateGuid, string hhTemplateName, int repetition = 0)
         {
-            string testID = "CalculationTests.SystematicHouseholdTemplateTest." + hhTemplateGuid + number;
+            string testID = "CalculationTests.SystematicHouseholdTemplateTestShort." + hhTemplateName + "." + repetition;
             HouseJobTestHelper.GenerateAndSimulateHHFromTemplate(testID, hhTemplateGuid, TestDuration.ThreeDays);
         }
 
@@ -875,46 +876,48 @@ namespace SimulationEngine.Tests {
         /// template, and simulates each household for one year.
         /// </summary>
         /// <param name="hhTemplateGuid">Guid of the template to test</param>
-        /// <param name="number">Repetition index of this test for the same template guid</param>
+        /// <param name="hhTemplateName">Name of the template to test</param>
+        /// <param name="repetition">Repetition of this test for the same template</param>
         /// <remarks>Each call of this method only handles a single household</remarks>
         [Theory]
         [MemberData(nameof(GetEachHHTemplate3Times))]
         [Trait(UnitTestCategories.Category, UnitTestCategories.HouseholdTemplateTest)]
-        public void SystematicHouseholdTemplateTestLong(StrGuid hhTemplateGuid, int number)
+        public void SystematicHouseholdTemplateTestLong(StrGuid hhTemplateGuid, string hhTemplateName, int repetition)
         {
-            string testID = "CalculationTests.SystematicHouseholdTemplateTest." + hhTemplateGuid + "." + number;
+            string testID = "CalculationTests.SystematicHouseholdTemplateTestLong." + hhTemplateName + "." + repetition;
             HouseJobTestHelper.GenerateAndSimulateHHFromTemplate(testID, hhTemplateGuid, TestDuration.TwelveMonths);
         }
 
         /// <summary>
-        /// Creates an enumerable of all household template guids, each one repeated 100 times.
+        /// Creates an enumerable of all household template guids and names, repeating each template 100 times.
         /// </summary>
-        /// <returns>An <see cref="IEnumerable{T}"/> of household template guids</returns>
+        /// <returns>An <see cref="IEnumerable{T}"/> of household template guids and names</returns>
         public static IEnumerable<object[]> GetEachHHTemplate100Times()
         {
             const string testname = "CalculationTests.GetEachHHTemplate100Times";
-            return CreateTemplateGuidList(testname, 100);
+            return CreateTemplateList(testname, 100);
         }
 
         /// <summary>
-        /// Creates an enumerable of all household template guids, each one repeated 3 times.
+        /// Creates an enumerable of all household template guids and names, repeating each template 3 times.
         /// </summary>
-        /// <returns>An <see cref="IEnumerable{T}"/> of household template guids</returns>
[…]
