### 2022-01-05 e1a06c3
Added check whether to use TestOutputHelper to avoid exceptions in TeamCity
