### 2022-01-05 1d5b131
Replace invalid characters when creating working directories with household template names
