### 2022-01-11 464e803
Invalid datetimes occur infrequently in template tests: save affected DBs to examine them later
