### 2022-01-11 e2e8d70
Fixed enabling idle mode in template tests
