### 2022-01-25 2cd2196
Added flexibility options for time-shifting energy demand to the UI
