+++ WpfApplication1/Presenters/Households/FlexbilityType.cs
@@ -1,0 +1,8 @@
+﻿//-----------------------------------------------------------------------
+
+namespace LoadProfileGenerator.Presenters.Households
+{
+    internal class FlexbilityType
+    {
+    }
+}
