### 2022-01-31 #3574 @fraukewiese
@noah80  Just a minor comment on the article: In the section Examples for previous usage, could you remove the brackets of the references since it is directly integrated in the sentence. Thus: "Harder et al. (2020) analyzed operating flexibility in electricity grids." instead of (Harder et al., 2020) analyzed operating flexibility in electricity grids.
