+++ VersionInfo.cs
@@ -1,3 +1,3 @@
-[assembly: AssemblyVersion("10.7.0.6")]
+[assembly: AssemblyVersion("10.8.0.1")]
 
-[assembly: AssemblyFileVersion("10.7.0.6")]
+[assembly: AssemblyFileVersion("10.8.0.1")]
