### 2022-03-03 7adf212
References to Enums in the PythonBindings are now generated with the correct type instead of 'str'
