### 2022-03-03 b9c8eb8
Update paper.md
fixed some reference formatting in JOSS paper
