### 2022-03-03 #3574 @kyleniemeyer
Hi @noah80, I just submitted a small PR to fix some reference formatting in the paper: https://github.com/FZJ-IEK3-VSA/LoadProfileGenerator/pull/15

Could you merge this? I can then accept the paper.
